LazyNode (dyck-typing Node) removed, integrated into Node
